pyproject.toml
[tool.poetry]
name = "remote-inky"
version = "0.1.0"
description = ""
authors = ["Axel <dev@absalon.is>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
numpy = "^1.20.2"
"rpi.gpio" = "^0.7.0"
spidev = "^3.5.0"

[tool.poetry.dev-dependencies]
isort = "^5.8.0"
black = "^20.8b1"
mypy = "^0.812"
flake8 = "^3.9.0"
flake8-builtins = "^1.5.3"
flake8-isort = "^4.0.0"
flake8-bugbear = "^21.4.3"
flake8-black = "^0.2.1"
flake8-comprehensions = "^3.4.0"
flake8-annotations = "^2.6.2"
ipdb = "^0.13.7"
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
